+++ pkg/deployment/informers.go
@@ -24,7 +24,7 @@ package deployment
 
 import (
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
+	crdv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
@@ -203,10 +203,10 @@ func (d *Deployment) listenForServiceEvents(stopCh <-chan struct{}) {
 func (d *Deployment) listenForCRDEvents(stopCh <-chan struct{}) {
 	rw := k8sutil.NewResourceWatcher(
 		d.deps.Log,
-		d.deps.KubeExtCli.ApiextensionsV1beta1().RESTClient(),
+		d.deps.KubeExtCli.ApiextensionsV1().RESTClient(),
 		"customresourcedefinitions",
 		"",
-		&v1beta1.CustomResourceDefinition{},
+		&crdv1.CustomResourceDefinition{},
 		cache.ResourceEventHandlerFuncs{
 			AddFunc: func(obj interface{}) {
 				d.triggerCRDInspection()